Make it work on Windows with Visual Studio.
Also, silence all Visual Studio warnings.

+++ Source/Game/Save.h
@@ -39,8 +39,8 @@ extern SaveData* const Save;
 #define offsetofSaveData(m) (offsetof(SaveData, m) + 0x100 - sizeof(void*) * ((sizeof(SaveData) / sizeof(void*)) + (sizeof(SaveData) % sizeof(void*) != 0)))
 
 // When accessing EEP-ROM to read/write save data, use these.
-#define WriteSave(m) WriteEeprom(offsetofSaveData(m), &Save->m, sizeof(Save->m));
-#define ReadSave(m) ReadEeprom(offsetofSaveData(m), &Save->m, sizeof(Save->m));
+#define WriteSave(m) WriteEeprom((uint8_t)offsetofSaveData(m), &Save->m, sizeof(Save->m));
+#define ReadSave(m) ReadEeprom((uint8_t)offsetofSaveData(m), &Save->m, sizeof(Save->m));
 
 extern uint32_t BestMasterSectionTimes[10];
 extern uint32_t BestTaDeathSectionTimes[10];
